// lib/interfaces.tsx
interface mouse {
  id: string;
  price: number | null;
  tracking_method: string | null;
  connection_type: string | null;
  max_dpi: number | null;
  hand_orientation: string | null;
  color: string | null;
  createdAt: string | null;
  updatedAt: string | null;
  manufacturer: string | null;
  name: string | null;
}